Version 20201127.1
https://www.gurux.fi/node/17294

+++ development/include/GXDLMSSFSKActiveInitiator.h
@@ -52,7 +52,7 @@ class CGXDLMSSFSKActiveInitiator : public CGXDLMSObject
     /**
      * MAC address of active initiator.
      */
-    int m_MacAddress;
+    uint16_t m_MacAddress;
     /**
      * L SAP selector of active initiator.
      */
